Move commands to a shared location
Had to convert to JS to silence Cyrpess TS errors
