+++ lib/cypress/commands.js
@@ -1,5 +1,3 @@
-/// <reference types="cypress" />
-
 // Asset that el has dimensions
 Cypress.Commands.add('hasDimensions',
   { prevSubject: true },
@@ -18,21 +16,18 @@ Cypress.Commands.add('isPlaying',
   cy.wrap(subject).should('have.prop', 'paused', false)
 })
 
-// Add Typescript support for custom commaands
-// https://docs.cypress.io/guides/tooling/typescript-support#Types-for-Custom-Commands
-export {};
-declare global {
-  namespace Cypress {
-    interface Chainable {
+// Check that a video is paused
+// https://glebbahmutov.com/blog/test-video-play/
+Cypress.Commands.add('isPaused',
+  { prevSubject: true },
+  (subject) => {
+  cy.wrap(subject).should('have.prop', 'paused', true)
+})
 
-      hasDimensions(
-        width: number,
-        height: number
-      ): Chainable<JQueryWithSelector>
-
-      isPlaying(): Chainable<void>
-    }
-  }
-}
-
-
+// Clear browser cache
+// https://stackoverflow.com/a/72945339/59160
+Cypress.Commands.add('clearCache', () => {
+  cy.wrap(Cypress.automation('remote:debugger:protocol', {
+    command: 'Network.clearBrowserCache',
+  }))
+})
