+++ packages/react/cypress/support/component.ts
@@ -1,5 +1,5 @@
 // Import commands.js using ES2015 syntax:
-import './commands'
+import '../../../../lib/cypress/commands'
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
